refactor(engine): delete the operator tier
kild had three baked-in ranks: operator, lead, worker. The operator was a
privileged session type the engine granted room-control tools to. In the
agent-centric model the human's agent CLI drives the REST API as an
ordinary client, so there is no tier to grant: kild ships no brain.

Removes the four operator tools, the KILD_OPERATOR capability in worker.ts
and server.ts, and the whole `kild operator` CLI group. fleetMemorySection
goes with them — it existed only to give operator sessions cross-project
memory and had no other caller.

engine-client.ts and rooms-status.ts move up out of operator/: neither is
an operator concern. rooms-status carries the cross-kild collision
computation, which is general observability that was only ever filed in
the wrong folder.

One test goes with its subject (fleetMemorySection); no other test failed
at any point. 259 tests pass, typecheck and lint clean.

pi-extension still references the removed capability. It is rebuilt
against the new shape in a later commit rather than patched twice.
